requirements.txt
Flask
Flask-WTF
gunicorn
flask-socketio
requests
psycopg2-binary
SQLAlchemy
Flask-SQLAlchemy
redis